% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FELI2.R
\name{FELI2}
\alias{FELI2}
\title{Meta-analysis for fixed effects - LI2}
\usage{
FELI2(info,alpha=0.05)
}
\arguments{
\item{info}{summary statistics}

\item{alpha}{significance level, default alpha=0.05}
}
\description{
Meta-analysis for fixed effects - LI2
}
\note{
format for summary statistics (point estimate 1, point estimate 2)
}
\examples{
study1 <- c(247.5, 178.0)
study2 <- c(270.0, 167.0)
study3 <- c(243.0, 186.3)
study4 <- c(215.2, 133.7)
study5 <- c(245.5, 171.0)
info <- rbind(study1,study2,study3,study4,study5)
FELI2(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean, Omer Ozturk
}
\keyword{Meta-analysis}
\keyword{robust}
