% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEWNPLI.R
\name{FEWNPLI}
\alias{FEWNPLI}
\title{Meta-analysis for fixed-effect - WNPLI}
\usage{
FEWNPLI(info,alpha=0.05)
}
\arguments{
\item{info}{summary statistics}

\item{alpha}{significance level, default alpha=0.05}
}
\description{
Meta-analysis for fixed-effect - WNPLI
}
\note{
format for summary statistics (shift, sample size 1, sample size 2)
}
\examples{
study1 <- c(69.5, 17, 16)
study2 <- c(103, 20, 30)
info <- rbind(study1,study2)
FEWNPLI(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean, Omer Ozturk
}
\keyword{Meta-analysis}
\keyword{robust}
