% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEWPS.R
\name{FEWPS}
\alias{FEWPS}
\title{Meta-analysis for fixed effects - two-sample Wilcoxon pooled}
\usage{
FEWPS(info,alpha=0.05)
}
\arguments{
\item{info}{summary statistics}

\item{alpha}{significance level, default alpha=0.05}
}
\description{
Meta-analysis for fixed effects - two-sample Wilcoxon pooled
}
\note{
format for summary statistics (shift, scale)
}
\examples{
study1 <- c(69.5, 11.65)
study2 <- c(103, 18.2)
info <- rbind(study1,study2)
FEWPS(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean, Omer Ozturk
}
\keyword{Meta-analysis}
\keyword{robust}
