% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2SHL12.R
\name{get2SHL12}
\alias{get2SHL12}
\title{Non-Pooled Variance Two-sample Hodges Lehmann estimate}
\usage{
get2SHL12(dat, alpha=0.05)
}
\arguments{
\item{dat}{full sample dataframe}

\item{alpha}{significance level}
}
\description{
Calculate two-sample Hodges Lehmann estimate with non-pooled variance
}
\examples{
x1 <- rnorm(15)
x2 <- rnorm(15)
dat <- data.frame(x1,x2)
get2SHL12(dat,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Hodges}
\keyword{Lehmann}
\keyword{estimate}
\keyword{statistics}
\keyword{summary}
\keyword{two-sample}
