% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfo.R
\name{getInfo}
\alias{getInfo}
\title{Summary statistics}
\usage{
getInfo(dat, alpha=0.05)
}
\arguments{
\item{dat}{full sample dataframe}

\item{alpha}{significance level for tau estimation}
}
\description{
Calculate summary statistics
}
\examples{
x1 <- rnorm(15)
x2 <- rnorm(15)
dat <- data.frame(x1,x2)
getInfo(dat,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean, Omer Ozturk
}
\keyword{Hodges}
\keyword{Lehmann}
\keyword{estimate}
\keyword{statistics}
\keyword{summary}
