% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khscale.R
\name{khscale}
\alias{khscale}
\title{Kloke-Yabes procedure}
\usage{
khscale(estmat,ssmat)
}
\arguments{
\item{estmat}{shift estimates matrix}

\item{ssmat}{sample size matrix}
}
\description{
Kloke-Yabes procedure
}
\note{
format for summary statistics (location)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Meta-analysis}
\keyword{robust}
