% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khscale2S.R
\name{khscale2S}
\alias{khscale2S}
\title{Kloke-Yabes procedure - two-sample}
\usage{
khscale2S(deltahat,mumat,ssmat)
}
\arguments{
\item{deltahat}{shift estimates matrix}

\item{mumat}{one-sample location estimates matrix}

\item{ssmat}{sample size matrix}
}
\description{
Kloke-Yabes procedure - two-sample
}
\note{
format for summary statistics (location)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Meta-analysis}
\keyword{robust}
