% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.forest.R
\name{meta.forest}
\alias{meta.forest}
\title{Meta-analysis Forest Plot}
\usage{
meta.forest(data, effect = "fixed", method='MMPS', fullsample = TRUE, xlab=c(""), pooled=FALSE)
}
\arguments{
\item{data}{input data}

\item{effect}{fixed effects or random effects}

\item{method}{Meta-analysis methods, default MMNP}

\item{fullsample}{data type. full sample or summary statistics for limited information}

\item{xlab}{label for the plot}

\item{pooled}{pooled variance or non-pooled variance}
}
\description{
Forest Plot over all studies
}
\examples{
meta.forest(BMI, effect = "fixed", method='HLNP', fullsample = TRUE)
}
\author{
Yanda Lang, Joseph McKean, Omer Ozturk
}
\keyword{Meta-analysis}
\keyword{robust}
