% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.qqplot.R
\name{meta.qqplot}
\alias{meta.qqplot}
\title{Meta-analysis q-q Plots}
\usage{
meta.qqplot(data, comparison = FALSE, effect = "fixed", residual = "MMPS")
}
\arguments{
\item{data}{input data}

\item{comparison}{indicating q-q plots for comparison over all studies or one q-q plot for the entire study; default comparison q-q plots}

\item{effect}{fixed effects or random effects}

\item{residual}{Meta-analysis methods, default MMNP}
}
\description{
q-q plots based on selected residuals
}
\examples{
sample1_treat <- rnorm(25,0,1)
sample1_con <- rnorm(25,0,1)
sample2_treat <- rnorm(25,1,1)
sample2_con <- rnorm(25,1,1)
dat <- data.frame(sample1_treat,sample1_con,sample2_treat,sample2_con)
meta.qqplot(dat, effect = "fixed", residual = "MMPS")
}
\author{
Yanda Lang, Joseph McKean, Omer Ozturk
}
\keyword{Meta-analysis}
\keyword{robust}
