% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaFE1S2T.R
\name{MetaFE1S2T}
\alias{MetaFE1S2T}
\title{Meta-analysis for fixed-effect non-pooled variance}
\usage{
MetaFE1S2T(info,alpha=0.05)
}
\arguments{
\item{info}{summary statistics}

\item{alpha}{significance level}
}
\description{
Meta-analysis for fixed-effect non-pooled variance
}
\note{
format for summary statistics (location, scale, sample size)
}
\examples{
study1_treat <- c(178.0, 109.5, 17)
study1_cont <- c(247.5, 113.8, 16)
study2_treat <- c(167.0, 162.6, 23)
study2_cont <- c(270.0, 93.8, 23)
info <- rbind(study1_treat,study1_cont,study2_treat,study2_cont)
MetaFE1S2T(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Meta-analysis}
\keyword{robust}
