% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaFE2SHL.R
\name{MetaFE2SHL}
\alias{MetaFE2SHL}
\title{Meta-analysis for fixed-effect - two-sample Hodges Lehmann}
\usage{
MetaFE2SHL(info,alpha=0.05)
}
\arguments{
\item{info}{summary statistics}

\item{alpha}{significance level}
}
\description{
Meta-analysis for fixed-effect - two-sample Hodges Lehmann
}
\note{
format for summary statistics (location, scale)
}
\examples{
study1 <- c(-80.0, 1243.9)
study2 <- c(-89.0, 1029.3)
info <- rbind(study1,study2)
MetaFE2SHL(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Meta-analysis}
\keyword{robust}
