% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaFEC2.R
\name{MetaFEC2}
\alias{MetaFEC2}
\title{Meta-analysis for fixed-effect - C2}
\usage{
MetaFEC2(info,alpha=0.05)
}
\arguments{
\item{info}{sample median}

\item{alpha}{significance level}
}
\description{
Meta-analysis for fixed-effect - C2
}
\note{
format for info (median)
}
\examples{
study1_treat <- c(178.0)
study1_cont <- c(247.5)
study2_treat <- c(167.0)
study2_cont <- c(270.0)
info <- rbind(study1_treat,study1_cont,study2_treat,study2_cont)
MetaFEC2(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Meta-analysis}
\keyword{robust}
