% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaFEC3.R
\name{MetaFEC3}
\alias{MetaFEC3}
\title{Meta-analysis for fixed-effect - C3}
\usage{
MetaFEC3(info,alpha=0.05)
}
\arguments{
\item{info}{sample median}

\item{alpha}{significance level}
}
\description{
Meta-analysis for fixed-effect - C3
}
\note{
format for info (median, sample size)
}
\examples{
study1_treat <- c(178.0, 17)
study1_cont <- c(247.5, 16)
study2_treat <- c(167.0, 23)
study2_cont <- c(270.0, 23)
info <- rbind(study1_treat,study1_cont,study2_treat,study2_cont)
MetaFEC3(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Meta-analysis}
\keyword{robust}
