% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaFED2.R
\name{MetaFED2}
\alias{MetaFED2}
\title{Meta-analysis for fixed-effect - D2}
\usage{
MetaFED2(info,alpha=0.05)
}
\arguments{
\item{info}{Hodges Lehmann estimates}

\item{alpha}{significance level}
}
\description{
Meta-analysis for fixed-effect - D2
}
\note{
format for info (Hodges Lehmann estimates)
}
\examples{
study1_treat <- c(178.0)
study1_cont <- c(247.5)
study2_treat <- c(167.0)
study2_cont <- c(270.0)
info <- rbind(study1_treat,study1_cont,study2_treat,study2_cont)
MetaFED2(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Meta-analysis}
\keyword{robust}
