% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaFED3.R
\name{MetaFED3}
\alias{MetaFED3}
\title{Meta-analysis for fixed-effect - D3}
\usage{
MetaFED3(info,alpha=0.05)
}
\arguments{
\item{info}{Hodges Lehmann estimates}

\item{alpha}{significance level}
}
\description{
Meta-analysis for fixed-effect - D3
}
\note{
format for info (Hodges Lehmann estimates, sample size)
}
\examples{
study1_treat <- c(178.0, 17)
study1_cont <- c(247.5, 16)
study2_treat <- c(167.0, 23)
study2_cont <- c(270.0, 23)
info <- rbind(study1_treat,study1_cont,study2_treat,study2_cont)
MetaFED3(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Meta-analysis}
\keyword{robust}
