% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaFEKY1S.R
\name{MetaFEKY1S}
\alias{MetaFEKY1S}
\title{Meta-analysis for fixed-effect - KY}
\usage{
MetaFEKY1S(info,alpha=0.05)
}
\arguments{
\item{info}{summary statistics}

\item{alpha}{significance level}
}
\description{
Meta-analysis for fixed-effect - KY
}
\note{
format for summary statistics (location,sample size)
}
\examples{
study1_treat <- c(178.0,17)
study1_cont <- c(247.5,16)
study2_treat <- c(167.0,23)
study2_cont <- c(270.0,23)
info <- rbind(study1_treat,study1_cont,study2_treat,study2_cont)
MetaFEKY1S(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Meta-analysis}
\keyword{robust}
