% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaREC4.R
\name{MetaREC4}
\alias{MetaREC4}
\title{Meta-analysis for random-effect - C4}
\usage{
MetaREC4(info,alpha=0.05)
}
\arguments{
\item{info}{median, standard deviation, and sampel size}

\item{alpha}{significance level}
}
\description{
Meta-analysis for random-effect - C4
}
\note{
format for info (median,sd,sample size)
}
\examples{
study1_treat <- c(178.0,198.9,17)
study1_cont <- c(247.5,179.2,16)
study2_treat <- c(167.0,118.8,23)
study2_cont <- c(270.0,90.8,23)
info <- rbind(study1_treat,study1_cont,study2_treat,study2_cont)
MetaREC4(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Meta-analysis}
\keyword{robust}
