% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaREKY1S.R
\name{MetaREKY1S}
\alias{MetaREKY1S}
\title{Meta-analysis for random-effect - KY}
\usage{
MetaREKY1S(info,alpha=0.05)
}
\arguments{
\item{info}{summary statistics}

\item{alpha}{significance level}
}
\description{
Meta-analysis for random-effect - KY
}
\note{
format for summary statistics (location,scale,sample size)
}
\examples{
study1_treat <- c(178.0, 109.5, 17)
study1_cont <- c(247.5, 113.8, 16)
study2_treat <- c(167.0, 162.6, 23)
study2_cont <- c(270.0, 93.8, 23)
info <- rbind(study1_treat,study1_cont,study2_treat,study2_cont)
MetaREKY1S(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Meta-analysis}
\keyword{robust}
