% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaREKY2S.R
\name{MetaREKY2S}
\alias{MetaREKY2S}
\title{Meta-analysis for random-effect - two-sample KY}
\usage{
MetaREKY2S(info,alpha=0.05)
}
\arguments{
\item{info}{shift estimates, scale estimates, location parameters, and sampel size}

\item{alpha}{significance level}
}
\description{
Meta-analysis for random-effect - two-sample KY
}
\note{
format for info (location,scale,location parameter1,location parameter2,sample size 1,sample size 2)
}
\examples{
study1_2SHL <- c(-80.0, 1243.9, 178.0, 247.5, 17, 16)
study2_2SHL <- c(-89.0, 1029.3, 167.0, 270.0, 23, 23)
info <- rbind(study1_2SHL,study2_2SHL)
MetaREKY2S(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean
}
\keyword{Meta-analysis}
\keyword{robust}
