% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBS2S.R
\name{REBS2S}
\alias{REBS2S}
\title{Meta-analysis for random effects - two sample bootstrap}
\usage{
REBS2S(info,alpha=0.05)
}
\arguments{
\item{info}{summary statistics}

\item{iteration}{number of iterations for bootstrap, default iteration=1000}

\item{alpha}{significance level, default alpha=0.05}
}
\description{
Meta-analysis for random effects - two sample bootstrap
}
\note{
format for summary statistics (shift, sample size 1, sample size 2)
}
\examples{
study1 <- c(69.5, 17, 16)
study2 <- c(103, 20, 30)
info <- rbind(study1,study2)
REBS2S(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean, Omer Ozturk
}
\keyword{Meta-analysis}
\keyword{robust}
