% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTWPS.R
\name{RESTWPS}
\alias{RESTWPS}
\title{Meta-analysis for random effects - two-sample Wilcoxon pooled variance}
\usage{
RESTWPS(info,alpha=0.05)
}
\arguments{
\item{info}{summary statistics}

\item{alpha}{significance level}
}
\description{
Meta-analysis for random effects - two-sample Wilcoxon pooled variance
}
\note{
format for summary statistics (shift, scale 1, scale 2, sample size 1, sample size 2)
}
\examples{
study1 <- c(69.5, 19.5, 13.8, 17, 16)
study2 <- c(103, 12.6, 9.8, 20, 30)
info <- rbind(study1,study2)
RESTWPS(info,alpha=0.05)
}
\author{
Yanda Lang, Joseph McKean, Omer Ozturk
}
\keyword{Meta-analysis}
\keyword{robust}
